+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## [In Development](https://github.com/ppfeufer/bootstrap-3-shortcodes-for-wordpress)
 ### Added
 - Shortcode ```[alert]```
+- Shortcode ```[badge]```
+- Shortcode ```[breadcrumb]``` and ```[breadcrumb-item]```